// packages/js-core/src/proto/polycentric.ts
export interface PublicKey {
  /**
   * @generated from protobuf field: uint64 key_type = 1
   */
  keyType: bigint;
  /**
   * @generated from protobuf field: bytes key = 2
   */
  key: Uint8Array;
}
/**
 * @generated from protobuf message polycentric.PrivateKey
 */
export interface PrivateKey {
  /**
   * @generated from protobuf field: uint64 key_type = 1
   */
  keyType: bigint;
  /**
   * @generated from protobuf field: bytes key = 2
   */
  key: Uint8Array;
}
// ...
class PublicKey$Type extends MessageType<PublicKey> {
  constructor() {
    super('polycentric.PublicKey', [
      {
        no: 1,
        name: 'key_type',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 2, name: 'key', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
    ]);
  }
  create(value?: PartialMessage<PublicKey>): PublicKey {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.keyType = 0n;
    message.key = new Uint8Array(0);
    if (value !== undefined)
      reflectionMergePartial<PublicKey>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: PublicKey,
  ): PublicKey {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* uint64 key_type */ 1:
          message.keyType = reader.uint64().toBigInt();
          break;
        case /* bytes key */ 2:
          message.key = reader.bytes();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: PublicKey,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* uint64 key_type = 1; */
    if (message.keyType !== 0n)
      writer.tag(1, WireType.Varint).uint64(message.keyType);
    /* bytes key = 2; */
    if (message.key.length)
      writer.tag(2, WireType.LengthDelimited).bytes(message.key);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message polycentric.PublicKey
 */
export const PublicKey = new PublicKey$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PrivateKey$Type extends MessageType<PrivateKey> {
  constructor() {
    super('polycentric.PrivateKey', [
      {
        no: 1,
        name: 'key_type',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 2, name: 'key', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
    ]);
  }
  create(value?: PartialMessage<PrivateKey>): PrivateKey {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.keyType = 0n;
    message.key = new Uint8Array(0);
    if (value !== undefined)
      reflectionMergePartial<PrivateKey>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: PrivateKey,
  ): PrivateKey {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* uint64 key_type */ 1:
          message.keyType = reader.uint64().toBigInt();
          break;
        case /* bytes key */ 2:
          message.key = reader.bytes();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: PrivateKey,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* uint64 key_type = 1; */
    if (message.keyType !== 0n)
      writer.tag(1, WireType.Varint).uint64(message.keyType);
    /* bytes key = 2; */
    if (message.key.length)
      writer.tag(2, WireType.LengthDelimited).bytes(message.key);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message polycentric.PrivateKey
 */
export const PrivateKey = new PrivateKey$Type();